requirements_laptop.txt
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
Babel
backcall
beautifulsoup4
bleach
bokeh
certifi
cffi
charset-normalizer
click
cloudpickle
colorama
colorcet
comm
constantly
contourpy
cycler
dask
datashader
datashape
debugpy
decorator
defusedxml
executing
fastjsonschema
filelock
fonttools
fqdn
fsspec
holoviews
idna
importlib-metadata
incremental
ipykernel
ipython
ipython-genutils
ipywidgets
isoduration
jedi
Jinja2
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
kiwisolver
linkify-it-py
llvmlite
locket
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdit-py-plugins
mdurl
mistune
monai
mpld3
mpmath
multipledispatch
mypy-extensions
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numba
numpy
opencv-python
overrides
packaging
pandas
pandocfilters
panel
param
parso
partd
pickleshare
Pillow
platformdirs
prometheus-client
prompt-toolkit
psutil
pure-eval
pyasn1
pycparser
pyct
Pygments
pynndescent
pyparsing
PyPDF2
pyrsistent
python-dateutil
python-json-logger
pytorch-ignite
pytz
pyviz_comms
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
rfc3339-validator
rfc3986-validator
roman
rpds-py
scikit-learn
scipy
Send2Trash
six
sniffio
soupsieve
sqlparse
stack-data
staintools
sympy
terminado
threadpoolctl
tinycss2
toolz
torch
tornado
tqdm
traitlets
txaio
types-python-dateutil
typing-extensions
tzdata
uc-micro-py
umap-learn
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
websockets
widgetsnbextension
xarray
xyzservices
zipp
